refactor: colocate announcementData with AnnouncementDialog
