+++ src/routes/home/components/announcementDialog/data/announcementData.js
@@ -5,7 +5,7 @@ export const announcementData = {
   activatorText: 'Recycler Fives are here to stay!',
   title: 'Keeping Recycler Fives',
   imgSrc: null,
-  displayCards: [Card.FIVE_OF_CLUBS, Card.FIVE_OF_HEARTS, Card.FIVE_OF_DIAMONDS, Card.FIVE_OF_SPADES],
+  displayCards: [Card.FIVE_OF_CLUBS, Card.FIVE_OF_CLUBS, Card.FIVE_OF_CLUBS, Card.FIVE_OF_CLUBS],
   announcementText: [
     {
       heading: 'Recycler Fives are here to stay',
