+++ tests/e2e/specs/out-of-game/navigation.spec.js
@@ -1,5 +1,5 @@
 import { playerOne } from '../../fixtures/userFixtures';
-import { announcementData } from '../../../../src/data/announcementData';
+import { announcementData } from '../../../../src/routes/home/components/announcementDialog/data/announcementData';
 
 function verifyAuthenticatedLinks() {
   // Navigate to Rules
